refactor(electron): decompose monolithic main.ts into modular architecture
- Extract environment setup (GC, userData, flags, smoke tests) to setup/environment.ts
- Extract IPC sender validation and permission lockdown to setup/security.ts
- Extract protocol handlers and webview CSP to setup/protocols.ts
- Extract app lifecycle handlers (second-instance, activate, etc.) to lifecycle/appLifecycle.ts
- Extract graceful shutdown logic to lifecycle/shutdown.ts
- Extract BrowserWindow creation and security hardening to window/createWindow.ts
- Extract service orchestration and post-launch setup to window/windowServices.ts
- Extract power monitor suspend/resume to window/powerMonitor.ts
- Add window/windowRef.ts as zero-import leaf module for shared mainWindow state
- Reduce main.ts from 1815 to 179 lines as thin orchestrator
