+++ electron/window/windowRef.ts
@@ -1,0 +1,11 @@
+import type { BrowserWindow } from "electron";
+
+let mainWindow: BrowserWindow | null = null;
+
+export function setMainWindow(win: BrowserWindow | null): void {
+  mainWindow = win;
+}
+
+export function getMainWindow(): BrowserWindow | null {
+  return mainWindow;
+}
